+++ changelog.md
@@ -5,6 +5,7 @@
   - added symbol validation for subscribe_to_market_data
   - refactored terminal state storage
   - fixed historical market data HTTP requests for symbols with special characters
+  - breaking change: removed updatePending, originalComment fields, added brokerComment field
 
 14.3.0
   - added copyFactoryResourceSlots field to make it possible specify resource slots for CopyFactory 2 application